Fix deleteCategory postman test

+++ ewm-service/src/main/java/ewm/service/category/CategoryServiceImpl.java
@@ -7,6 +7,7 @@ import ewm.exception.NotFoundException;
 import ewm.mapper.category.CategoryMapper;
 import ewm.model.category.Category;
 import ewm.repository.category.CategoryRepository;
+import ewm.repository.event.EventRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -18,14 +19,14 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final EventRepository eventRepository;
 
     @Override
-    @Transactional
     public CategoryDto createCategory(NewCategoryDto newCategoryDto) {
         if (categoryRepository.existsByName(newCategoryDto.getName())) {
             throw new ConflictException("Category name already exists");
@@ -36,19 +37,16 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    @Transactional
     public void deleteCategory(Long catId) {
         Category category = getCategoryEntity(catId);
-//  Дописать после реализации eventRepository
-//        boolean hasEvents = eventRepository.existsByCategoryId(catId);
-//        if (hasEvents) {
-//            throw new ConflictException("The category is not empty");
-//        }
+        boolean hasEvents = eventRepository.existsByCategory_Id(catId);
+        if (hasEvents) {
+            throw new ConflictException("The category is not empty");
+        }
         categoryRepository.delete(category);
     }
 
     @Override
-    @Transactional
     public CategoryDto updateCategory(Long catId, CategoryDto categoryDto) {
         Category existingCategory = getCategoryEntity(catId);
         if (categoryDto.getName() != null && !categoryDto.getName().isBlank()) {
@@ -59,6 +57,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<CategoryDto> getCategories(Integer from, Integer size) {
         Pageable pageable = PageRequest.of(from / size, size);
         List<Category> categories = categoryRepository.findAll(pageable).getContent();
@@ -68,6 +67,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public CategoryDto getCategoryById(Long catId) {
         Category category = getCategoryEntity(catId);
         return categoryMapper.toDto(category);
